AI Workflow Functions/clear_tables/requirements.txt
functions-framework
cloud-sql-python-connector[pg8000]
sqlalchemy
google-cloud-secret-manager
